+++ src/app/screens/CreatePassword.test.tsx
@@ -7,7 +7,8 @@ import configureStore from 'redux-mock-store';
 import type { ApplicationState } from '@app/store';
 import { createPassword } from '@common/store';
 import { translateRaw } from '@common/translate';
-import { CreatePassword } from '@screens';
+
+import { CreatePassword } from './CreatePassword';
 
 jest.mock('@bridge', () => ({
   ipcBridgeRenderer: {
